// types/TabulatorTypes.ts
declare namespace Tabulator {
	interface Options
		extends OptionsGeneral,
			OptionsMenu,
			OptionsHistory,
			OptionsLocale,
			OptionsDownload,
			OptionsColumns,
			OptionsRows,
			OptionsData,
			OptionsSorting,
			OptionsFiltering,
			OptionsRowGrouping,
			OptionsPagination,
			OptionsPersistentConfiguration,
			OptionsClipboard,
			OptionsDataTree,
			OptionsCell,
			OptionsCells,
			OptionsDebug,
			OptionsHTML {}

	interface OptionsDebug {
		invalidOptionWarning?: boolean;
		debugInvalidOptions?: boolean;
	}
// ...
	interface OptionsPagination {
		pagination?: boolean;
		paginationMode?: SortMode;
		/** Set the number of rows in each page */
		paginationSize?: number | undefined;
		/**  Setting this option to true will cause Tabulator to create a list of page size options, that are multiples of the current page size. In the example below, the list will have the values of 5, 10, 15 and 20.

    When using the page size selector like this, if you use the setPageSize function to set the page size to a value not in the list, the list will be regenerated using the new page size as the starting valuer    */
		paginationSizeSelector?: true | number[] | any[] | undefined;
		/**  By default the pagination controls are added to the footer of the table. If you wish the controls to be created in another element pass a DOM node or a CSS selector for that element to the paginationElement option.*/
		paginationElement?: HTMLElement | string | undefined;
		/** Lookup list to link expected data feilds from the server to their function
		 * default* {
            "current_page":"current_page",
            "last_page":"last_page",
            "data":"data",
            }
		 *
		 */
		dataReceiveParams?: Record<string, string> | undefined;
		/** Lookup list to link fields expected by the server to their function* default:* {
        "page":"page",
        "size":"size",
        "sorters":"sorters",
        "filters":"filters",
        }
		 */
		dataSendParams?: Record<string, string> | undefined;
		/** When using the addRow function on a paginated table, rows will be added relative to the current page (ie to the top or bottom of the current page), with overflowing rows being shifted onto the next page.

      If you would prefer rows to be added relative to the table (firs/last page) then you can use the paginationAddRow option. it can take one of two values:

      page - add rows relative to current page (default)
      table - add rows relative to the table */
		paginationAddRow?: 'table' | 'page' | undefined;
		/**  The number of pagination page buttons shown in the footer using the paginationButtonCount option. By default this has a value of 5.*/
		paginationButtonCount?: number | undefined;
		/** Specify that a specific page should be loaded when the table first load */
		paginationInitialPage?: number | undefined;
	}

	type GroupArg = string | string[] | ((data: any) => any);

	interface OptionsRowGrouping {
		/** String/function to select field to group rows by     */
		groupBy?: GroupArg | undefined;
		/** By default Tabulator will create groups for rows based on the values contained in the row data. if you want to explicitly define which field values groups should be created for at each level, you can use the groupValues option.

    This option takes an array of value arrays, each item in the first array should be a list of acceptable field values for groups at that level     */
		groupValues?: GroupValuesArg | undefined;

		/** You can use the setGroupHeader function to change the header generation function for each group. This function has one argument and takes the same values as passed to the groupHeader setup option.     */
		groupHeader?:
			| ((value: any, count: number, data: any, group: GroupComponent) => string)
			| Array<(value: any, count: number, data: any) => string>
			| undefined;

		/**When printing you may want to apply a different group header from the one usualy used in the table. You can now do this using the groupHeaderPrint table option, which takes the same inputs as the standard groupHeader property. */
		groupHeaderPrint?:
			| ((value: any, count: number, data: any, group: GroupComponent) => string)
			| Array<(value: any, count: number, data: any) => string>
			| undefined;

		/** You can set the default open state of groups using the groupStartOpen property* * This can take one of three possible values:

        true - all groups start open (default value)
        false - all groups start closed
        function() - a callback to decide if a group should start open or closed
        Group Open Function
        If you want to decide on a group by group basis which should start open or closed then you can pass a function to the groupStartOpen property. This should return true if the group should start open or false if the group should start closed.
		 */
		groupStartOpen?:
			| boolean
			| ((value: any, count: number, data: any, group: GroupComponent) => boolean)
			| undefined;

		/** By default Tabulator allows users to toggle a group open or closed by clicking on the arrow icon in the left of the group header. If you would prefer a different behaviour you can use the groupToggleElement option to choose a different option:* * The option can take one of three values:
      arrow - togggle group on arrow element click
      header - toggle group on click anywhere on the group header element
      false - prevent clicking anywhere in the group toggling the group
		 */
		groupToggleElement?: 'arrow' | 'header' | false | undefined;

		/** show/hide column calculations when group is closed     */
		groupClosedShowCalcs?: boolean | undefined;

		/** The dataGrouping callback is triggered whenever a data grouping event occurs, before grouping happens. */
		dataGrouping?: (() => void) | undefined;
		/** The dataGrouping callback is triggered whenever a data grouping event occurs, after grouping happens. */
		dataGrouped?: (() => void) | undefined;
		/** The groupVisibilityChanged callback is triggered whenever a group changes between hidden and visible states. */
		groupVisibilityChanged?: ((group: GroupComponent, visible: boolean) => void) | undefined;

		/** The groupClick callback is triggered when a user clicks on a group header. */
		groupClick?: GroupEventCallback | undefined;
		/** The groupDblClick callback is triggered when a user double clicks on a group header. */
		groupDblClick?: GroupEventCallback | undefined;
		/** The groupContext callback is triggered when a user right clicks on a group header.

    If you want to prevent the browsers context menu being triggered in this event you will need to include the preventDefault() function in your callback. */
		groupContext?: GroupEventCallback | undefined;
		/** The groupTap callback is triggered when a user taps on a group header on a touch display. */
		groupTap?: GroupEventCallback | undefined;
		/** The groupDblTap callback is triggered when a user taps on a group header on a touch display twice in under 300ms. */
		groupDblTap?: GroupEventCallback | undefined;
		/** The groupTapHold callback is triggered when a user taps on a group header on a touch display and holds their finger down for over 1 second */
		groupTapHold?: GroupEventCallback | undefined;
		groupUpdateOnCellEdit?: boolean | undefined;
	}

	interface Filter {
		field: string;
		type: FilterType;
		value: any;
	}

	interface FilterParams {
		separator?: string | undefined;
		matchAll?: boolean | undefined;
	}
	type FilterFunction = (field: string, type: FilterType, value: any, filterParams?: FilterParams) => void;

	interface OptionsFiltering {
		/** Array of filters to be applied on load. */
		initialFilter?: Filter[] | undefined;

		/** array of initial values for header filters. */
		initialHeaderFilter?: Array<Pick<Filter, 'field' | 'value'>> | undefined;

		/** The dataFiltering callback is triggered whenever a filter event occurs, before the filter happens. */
		dataFiltering?: ((filters: Filter[]) => void) | undefined;
		/** The dataFiltered callback is triggered after the table dataset is filtered. */
		dataFiltered?: ((filters: Filter[], rows: RowComponent[]) => void) | undefined;

		/**When using real time header filtering, Tabulator will wait 300 miliseconds after a keystroke before triggering the filter. You can customise this delay by using the headerFilterLiveFilterDelay table setup option */
		headerFilterLiveFilterDelay?: number | undefined;
	}
	interface OptionsSorting {
		/** Array of sorters to be applied on load.     */
		initialSort?: Sorter[] | undefined;

		/** reverse the order that multiple sorters are applied to the table.     */
		sortOrderReverse?: boolean | undefined;
	}
// ...
	interface OptionsMenu {
		rowContextMenu?: RowContextMenuSignature | undefined;
		rowClickMenu?: RowContextMenuSignature | undefined;
		groupClickMenu?: GroupContextMenuSignature | undefined;

		groupContextMenu?: Array<MenuObject<GroupComponent>> | undefined;
	}
// ...
	interface OptionsDownload {
		/** The downloadReady callback allows you to intercept the download file data before the users is prompted to save the file.

    In order for the download to proceed the downloadReady callback is expected to return a blob of file to be downloaded.

    If you would prefer to abort the download you can return false from this callback. This could be useful for example if you want to send the created file to a server via ajax rather than allowing the user to download the file. */
		downloadReady?: ((fileContents: any, blob: any) => any) | undefined;

		/** The downloadComplete callback is triggered when the user has been prompted to download the file. */
		downloadComplete?: (() => void) | undefined;

		/** By default Tabulator includes column headers, row groups and column calculations in the download output.

    You can choose to remove column headers groups, row groups or column calculations from the output data by setting the values in the downloadConfig option in the table definition: */

		downloadConfig?: AddditionalExportOptions | undefined;

		/**By deafault, only the active rows (rows that have passed filtering) will be included in the download the downloadRowRange option takes a Row Range Lookup value and allows you to choose which rows are included in the download output */
		downloadRowRange?: RowRangeLookup | undefined;
	}
// ...
	interface OptionsLocale {
		/** You can set the current local in one of two ways. If you want to set it when the table is created, simply include the locale option in your Tabulator constructor. You can either pass in a string matching one of the language options you have defined, or pass in the boolean true which will cause Tabulator to auto-detect the browsers language settings from the navigator.language object. */
		locale?: boolean | string | undefined;

		/** You can store as many languages as you like, creating an object inside the langs object with a property of the locale code for that language. A list of locale codes can be found here.

    At present there are three parts of the table that can be localised, the column headers, the header filter placeholder text and the pagination buttons. To localize the pagination buttons, create a pagination property inside your language object and give it the properties outlined below.

    If you wish you can also localize column titles by adding a columns property to your language object. You should store a property of the field name of the column you wish to change, with a value of its title. Any fields that match this will use this title instead of the one provided by the column definition array. */
		langs?: any;

		/** When a localization event has occurred , the localized callback will triggered, passing the current locale code and language object: */
		localized?: ((locale: string, lang: any) => void) | undefined;
	}

	type HistoryAction = 'cellEdit' | 'rowAdd' | 'rowDelete' | 'rowMoved';
	interface OptionsHistory {
		/** Enable user interaction history functionality     */
		history?: boolean | undefined;

		/** The historyUndo event is triggered when the undo action is triggered. */
		historyUndo?: ((action: HistoryAction, component: CellComponent | RowComponent, data: any) => void) | undefined;
		/** The historyRedo event is triggered when the redo action is triggered. */
		historyRedo?: ((action: HistoryAction, component: CellComponent | RowComponent, data: any) => void) | undefined;
	}
// ...
}